python/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
